+++ packages/kit/test/apps/basics/src/routes/routing/hashes/pagestate/+page.svelte
@@ -1,5 +1,5 @@
 <script>
-	import { page } from '$app/stores';
+	import { page } from '$app/state';
 	import { onMount } from 'svelte';
 
 	/** @type {string} */
@@ -15,8 +15,8 @@
 <svelte:window on:hashchange={set_hash} />
 
 <h1 id="window-hash">{hash}</h1>
-<h1 id="page-url-hash">{$page.url.hash}</h1>
+<h1 id="page-url-hash">{page.url.hash}</h1>
 
 <a href="#target">Nav to hash</a>
-<a href="/routing/hashes/pagestore">Nav to page</a>
+<a href="/routing/hashes/pagestate">Nav to page</a>
 <div id="target">Target</div>