+++ playgrounds/basic/src/routes/+page.svelte
@@ -7,3 +7,9 @@
 <h1>Welcome to SvelteKit</h1>
 
 2 + 2 = {data.sum}
+
+<h2>Pages:</h2>
+
+<ul>
+	<li><a href="images">images</a></li>
+</ul>